# fwber-backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "FWBer API Documentation",
        "description": "Comprehensive API documentation for FWBer - A proximity-based dating platform with real-time features, AI-powered matching, and privacy-first design.",
        "contact": {
            "name": "FWBer Team",
            "email": "support@fwber.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/api",
            "description": "Local Development Server"
        },
        {
            "url": "https://api.fwber.com/api",
            "description": "Production Server"
        }
    ],
    "paths": {
...
        "/location": {
            "get": {
                "tags": [
                    "Location"
                ],
                "summary": "Get current user's location",
                "description": "Get current user's location",
                "operationId": "76e8a81981e2df10a6369fd7ff2aa02d",
                "responses": {
                    "200": {
                        "description": "Location"
                    },
                    "404": {
                        "description": "No location data"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Location"
                ],
                "summary": "Update current location",
                "description": "Updates the authenticated user's current location and privacy level.",
                "operationId": "70594e85005ccdebfc189822225e0acf",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "latitude",
                                    "longitude"
                                ],
                                "properties": {
                                    "latitude": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 37.7749
                                    },
                                    "longitude": {
                                        "type": "number",
                                        "format": "float",
                                        "example": -122.4194
                                    },
                                    "accuracy": {
                                        "type": "number",
                                        "example": 12.5
                                    },
                                    "heading": {
                                        "type": "number",
                                        "example": 140
                                    },
                                    "speed": {
                                        "type": "number",
                                        "example": 1.2
                                    },
                                    "altitude": {
                                        "type": "number",
                                        "example": 20
                                    },
                                    "privacy_level": {
                                        "type": "string",
                                        "enum": [
                                            "public",
                                            "friends",
                                            "private"
                                        ]
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Updated"
                    },
                    "422": {
                        "description": "Validation failed"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Location"
                ],
                "summary": "Clear current location",
                "description": "Clear location history (set inactive)",
                "operationId": "82875a613c6ad6a45dc6428f1c8ddfd8",
                "responses": {
                    "200": {
                        "description": "Cleared"
                    },
                    "404": {
                        "description": "No location data"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/location/nearby": {
            "get": {
                "tags": [
                    "Location"
                ],
                "summary": "Find nearby users",
                "description": "Finds nearby users within a radius considering privacy settings.",
                "operationId": "f5b585d4a0ba054d3f29ffc036b6be86",
                "parameters": [
                    {
                        "name": "latitude",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "number",
                            "format": "float"
                        }
                    },
                    {
                        "name": "longitude",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "number",
                            "format": "float"
                        }
                    },
                    {
                        "name": "radius",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "maximum": 10000,
                            "minimum": 100
                        }
                    },
                    {
                        "name": "limit",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "maximum": 100,
                            "minimum": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Nearby users list"
                    },
                    "422": {
                        "description": "Validation failed"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/location/privacy": {
            "put": {
                "tags": [
                    "Location"
                ],
                "summary": "Update location privacy",
                "description": "Update location privacy settings",
                "operationId": "e6842cc8f113f67517f3e9c4e9244d6d",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "privacy_level"
                                ],
                                "properties": {
                                    "privacy_level": {
                                        "type": "string",
                                        "enum": [
                                            "public",
                                            "friends",
                                            "private"
                                        ]
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Updated"
                    },
                    "404": {
                        "description": "No location data"
                    },
                    "422": {
                        "description": "Validation failed"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
...
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Enter your bearer token in the format: Bearer {token}",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}